chore: Revert plan client to use along with hierarchy
- add common types in doppler types
- implement DopplerPlanClient hierarchy
- remove unused tests
