+++ src/services/pure-di.tsx
@@ -10,6 +10,7 @@ import { DopplerApiClient, HttpDopplerApiClient } from './doppler-api-client';
 import { DopplerSitesClient, HttpDopplerSitesClient } from './doppler-sites-client';
 import { IpinfoClient, HttpIpinfoClient } from './ipinfo-client';
 import { ExperimentalFeatures } from './experimental-features';
+import { DopplerPlanClient } from './doppler-plan-client';
 
 interface AppConfiguration {
   dopplerLegacyUrl: string;
@@ -42,6 +43,7 @@ export interface AppServices {
   experimentalFeatures: ExperimentalFeatures;
   dopplerApiClient: DopplerApiClient;
   ipinfoClient: IpinfoClient;
+  dopplerPlanClient: DopplerPlanClient;
 }
 
 /**
@@ -113,6 +115,13 @@ export class AppCompositionRoot implements AppServices {
           baseUrl: this.appConfiguration.dopplerLegacyUrl,
           window: this.window,
         }),
+    );
+  }
+
+  get dopplerPlanClient() {
+    return this.singleton(
+      'dopplerPlanClient',
+      () => new DopplerPlanClient({ dopplerLegacyClient: this.dopplerLegacyClient }),
     );
   }
